Various minor bugfixes - common/loader.c: rebuilt include/map.h: Change width,height,enter_x,enter_y to be 16 bit instead    of 8 bits. lib/treasures: change skull treasure to tooth instead of skull_tooth since    the later is not currently in the distributed archetypes file server/monster.c: Optimize some code, remove some other that is not used. server/plugins.c: Fix un initialized value in initPlugins that would    cause crash at startup if plugin directory does not exist. MSW 2001-11-13
git-svn-id: svn://svn.code.sf.net/p/crossfire/code/trunk/crossfire@1414 282e977c-c81d-0410-88c4-b93c2d0d6712

+++ include/map.h
@@ -162,9 +162,9 @@ typedef struct mapdef {
     uint16 difficulty;	/* What level the player should be to play here */
 
     uint8 darkness;    	/* indicates level of darkness of map */
-    uint8 width,height;	/* Width and height of map. */
-    uint8 enter_x;	/* enter_x and enter_y are default entrance location */
-    uint8 enter_y;	/* on the map if none are set in the exit */
+    uint16 width,height;	/* Width and height of map. */
+    uint16 enter_x;	/* enter_x and enter_y are default entrance location */
+    uint16 enter_y;	/* on the map if none are set in the exit */
     oblinkpt *buttons;	/* Linked list of linked lists of buttons */
     MapSpace	*spaces;    /* Array of spaces on this map */
     char    *msg;	/* Message map creator may have left */
